+++ CHANGELOG.md
@@ -13,6 +13,26 @@
 
 ### Security
 
+## [0.4.3] - 2024-07-31
+### Deprecated
+- The ability for prefab instances to be selected during generation has been deprecated due to an error
+
+### Fixed
+- ArgumentOutOfRangeException
+
+## [0.4.2] - 2024-07-31
+### Changed
+- add changelogurl to package.json
+- prefab instance is now selected when generated
+- The operation is now performed only on object with skinnedmeshrender
+
+### Fixed
+- ArgumentException: Can't save part of a Prefab instance as a Prefab
+- operations could be performed on prefab assets.
+- prefab instance was placed in different scene
+- skinnedMeshRenderer.bones was being called too much
+- namespace is not used
+
 ## [0.4.1] - 2024-05-18
 ### Changed
 - Change log message to appropriate format